Refactor response mapper implementation
